requirements.txt
asgiref==3.2.3
cachetools==4.0.0
certifi==2019.11.28
chardet==3.0.4
defusedxml==0.6.0
Django==3.0.3
django-crispy-forms==1.8.1
djangorestframework==3.11.0
google-api-python-client==1.7.11
google-auth==1.11.0
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.4.1
httplib2==0.17.0
idna==2.8
oauth2client==4.1.3
oauthlib==3.1.0
pkg-resources==0.0.0
psycopg2==2.8.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyJWT==1.7.1
pyspark==3.0.0.dev0
python3-openid==3.1.0
pytz==2019.3
requests==2.22.0
requests-oauthlib==1.3.0
rsa==4.0
six==1.14.0
social-auth-app-django==3.1.0
social-auth-core==3.2.0
sqlparse==0.3.0
uritemplate==3.0.1
urllib3==1.25.8
